adedSmarts

+++ Shop_back/Controllers/Items/SmartsController.cs
@@ -1,8 +1,9 @@
-﻿using Microsoft.AspNetCore.Mvc;
-using Shop_back.Contracts.Request.items;
+﻿using FluentValidation;
+using Microsoft.AspNetCore.Mvc;
+using Shop_back.Contracts.Request.items.Smart;
 using Shop_back.Contracts.Response.Items;
 using Shop_back.Core.Abstractions.Items.Smarts;
-using Shop_back.Core.Models.Items;
+using Shop_back.Core.Models.Items.Smart;
 
 namespace Shop_back.Controllers.Items
 {
@@ -11,40 +12,61 @@ namespace Shop_back.Controllers.Items
     public class SmartsController : ControllerBase
     {
         private readonly ISmartsService _service;
+        private readonly IValidator<CreateSmartRequest> _validator;
 
-        public SmartsController(ISmartsService service)
+        public SmartsController(ISmartsService service, IValidator<CreateSmartRequest> validator)
         {
             _service = service;
+            _validator = validator;
         }
         [HttpGet]
         public async Task<ActionResult<List<GetAllSmartsResponse>>> GetSmarts()
         {
             var smarts = await _service.GetAllSmarts();
-            var response = smarts.Select(s => new GetAllSmartsResponse(s.Id, s.Title, s.Description, s.Price));
+            var response = smarts.Select(s => new GetAllSmartsResponse(s.Id, s.Title, s.Description, s.Variants));
             return Ok(response);
         }
         [HttpPost]
         public async Task<ActionResult<Guid>> CreateSmart([FromBody] CreateSmartRequest req)
         {
-            var (smart, error) = Smart.Create(Guid.NewGuid(), req.Title, req.Description, req.Price);
-            if (!string.IsNullOrWhiteSpace(error)) return BadRequest(error);
+            var validationResult = await _validator.ValidateAsync(req);
+
+            if (!validationResult.IsValid)
+            {
+                var errors = validationResult.Errors
+                    .Select(e => e.ErrorMessage)
+                    .ToList();
+
+                return BadRequest(new { errors });
+            }
+            var options = req.Options
+                .Select(o => new SmartVariantOptions(
+                    o.Stock,
+                    o.Color,
+                    o.Memory,
+                    o.Storage,
+                    o.Price,
+                    o.Discount
+                ))
+                .ToArray();
+
+            var smart = new Smart(req.Title, req.Description, options);
+
             var smartId = await _service.CreateSmart(smart);
+
             return Ok(smartId);
         }
         [HttpPut("{id:guid}")]
         public async Task<ActionResult<Guid>> UpdateSmart(Guid id, [FromBody] CreateSmartRequest req)
         {
-            string titleError = Smart.ValidateFeld(req.Title, "Title");
-            string descError = Smart.ValidateFeld(req.Description, "Description");
-            string priceError = Smart.ValidatePrice(req.Price);
-            string error = string.Join("\n", titleError, descError, priceError);
-            if (!string.IsNullOrWhiteSpace(error)) return BadRequest(error);
-            var smartId = await _service.UpdateSmart(id, req.Title, req.Description, req.Price);
+            var smartId = await _service.UpdateSmart(id, req.Title, req.Description);
             return Ok(smartId);
         }
-        [HttpDelete("{id:guid}")] 
+        [HttpDelete("{id:guid}")]
         public async Task<ActionResult<Guid>> DeleteSmart(Guid id)
         {
+            var smart = await _service.GetSmartById(id);
+            if (smart == null) return NotFound($"smart with id {id} not found");
             var smartId = await _service.DeleteSmart(id);
             return Ok(smartId);
         }
